requirements.txt
django
djangorestframework
psycopg2-binary
sqlparse
asgiref
pymilvus==2.3.6
sentence-transformers
cffi
retry
pydantic
beautifulsoup4
nltk
